move the concrete tests to integrationtests.scala
