+++ spray-client/src/it/scala/com/heroku/platform/api/client/spray/AppSpec.scala
@@ -39,8 +39,3 @@ abstract class AppSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpec(
 
 }
 
-
-class SprayAppSpec extends AppSpec(SprayJsonBoilerplate) with SprayApiSpec
-
-
-class PlayAppSpec extends AppSpec(PlayJsonBoilerplate) with SprayApiSpec
